+++ game/unit/modular.cpp
@@ -587,7 +587,7 @@ auto mark::unit::modular::collide(const segment_t& ray) -> std::optional<
 	return { { std::ref(*damageable), *min } };
 }
 
-auto mark::unit::modular::collide(vector<double> center, float radius)
+auto mark::unit::modular::collide(vector<double> center, double radius)
 	-> std::vector<std::reference_wrapper<interface::damageable>>
 {
 	std::unordered_set<interface::damageable*> out;
