+++ game/world.cpp
@@ -301,7 +301,7 @@ auto mark::world::collide(const segment_t& ray) -> collide_result
 	return { move(collisions), map_collision };
 }
 
-auto mark::world::collide(vector<double> center, float radius)
+auto mark::world::collide(vector<double> center, double radius)
 	-> std::vector<std::reference_wrapper<interface::damageable>>
 {
 	std::vector<std::reference_wrapper<interface::damageable>> out;
